advisor-agent/backend/requirements.txt
fastapi
uvicorn[standard]
redis
python-dotenv
pydantic
prometheus-fastapi-instrumentator
requests
pytesseract
pdf2image
Pillow
psycopg2-binary
supabase
jira
python-jose
python-multipart
python-slugify
pdfplumber
